Refactor utilities.
