+++ src/libfuncs/bounded_int.rs
@@ -2,8 +2,10 @@
 
 use super::LibfuncHelper;
 use crate::{
-    block_ext::BlockExt, error::Result, metadata::MetadataStorage, types::TypeBuilder,
-    utils::RangeExt,
+    error::Result,
+    metadata::MetadataStorage,
+    types::TypeBuilder,
+    utils::{BlockExt, RangeExt},
 };
 use cairo_lang_sierra::{
     extensions::{
