+++ src/libfuncs/felt252_dict.rs
@@ -1,10 +1,10 @@
 //! # `Felt` dictionary libfuncs
 
 use super::LibfuncHelper;
-use crate::block_ext::BlockExt;
 use crate::{
     error::Result,
     metadata::{runtime_bindings::RuntimeBindingsMeta, MetadataStorage},
+    utils::BlockExt,
 };
 use cairo_lang_sierra::{
     extensions::{
