+++ src/libfuncs/struct.rs
@@ -1,9 +1,11 @@
 //! # Struct-related libfuncs
 
 use super::LibfuncHelper;
-use crate::block_ext::BlockExt;
 use crate::{
-    error::Result, metadata::MetadataStorage, types::TypeBuilder, utils::ProgramRegistryExt,
+    error::Result,
+    metadata::MetadataStorage,
+    types::TypeBuilder,
+    utils::{BlockExt, ProgramRegistryExt},
 };
 use cairo_lang_sierra::{
     extensions::{
