+++ src/utils.rs
@@ -1,23 +1,20 @@
 //! # Various utilities
 
-use crate::{metadata::MetadataStorage, types::TypeBuilder, OptLevel};
+pub(crate) use self::{
+    block_ext::BlockExt, program_registry_ext::ProgramRegistryExt, range_ext::RangeExt,
+};
+use crate::{metadata::MetadataStorage, OptLevel};
 use cairo_lang_compiler::CompilerConfig;
 use cairo_lang_sierra::{
-    extensions::{
-        core::{CoreLibfunc, CoreType},
-        utils::Range,
-    },
-    ids::{ConcreteTypeId, FunctionId},
+    ids::FunctionId,
     program::{GenFunction, Program, StatementIdx},
-    program_registry::ProgramRegistry,
 };
 use melior::{
-    ir::{Module, Type},
+    ir::Module,
     pass::{self, PassManager},
     Context, Error, ExecutionEngine,
 };
 use num_bigint::{BigInt, BigUint, Sign};
-use num_traits::One;
 use std::sync::LazyLock;
 use std::{
     alloc::Layout,
@@ -29,6 +26,10 @@ use std::{
     sync::Arc,
 };
 use thiserror::Error;
+
+mod block_ext;
+mod program_registry_ext;
+mod range_ext;
 
 #[cfg(target_os = "macos")]
 pub const SHARED_LIBRARY_EXT: &str = "dylib";
@@ -77,17 +78,10 @@ pub fn get_integer_layout(width: u32) -> Layout {
     } else if width <= 64 {
         Layout::new::<u64>()
     } else if width <= 128 {
-        #[cfg(not(target_arch = "x86_64"))]
-        {
-            Layout::new::<u128>()
-        }
-        #[cfg(target_arch = "x86_64")]
-        {
-            Layout::new::<u128>().align_to(16).unwrap()
-        }
+        Layout::new::<u128>()
     } else {
-        let width = (width as usize).next_multiple_of(8);
-        Layout::from_size_align(width >> 3, 16).expect("align should be power of two")
+        // According to the docs this should never return an error.
+        Layout::from_size_align((width as usize).next_multiple_of(8) >> 3, 16).unwrap()
     }
 }
 
@@ -457,100 +451,6 @@ pub fn layout_repeat(layout: &Layout, n: usize) -> Result<(Layout, usize), Layou
     // The safe constructor is called here to enforce the isize size limit.
     let layout = Layout::from_size_align(alloc_size, layout.align()).map_err(|_| LayoutError)?;
     Ok((layout, padded_size))
-}
-
-pub trait ProgramRegistryExt {
-    fn build_type<'ctx>(
-        &self,
-        context: &'ctx Context,
-        module: &Module<'ctx>,
-        registry: &ProgramRegistry<CoreType, CoreLibfunc>,
-        metadata: &mut MetadataStorage,
-        id: &ConcreteTypeId,
-    ) -> Result<Type<'ctx>, super::error::Error>;
-
-    fn build_type_with_layout<'ctx>(
-        &self,
-        context: &'ctx Context,
-        module: &Module<'ctx>,
-        registry: &ProgramRegistry<CoreType, CoreLibfunc>,
-        metadata: &mut MetadataStorage,
-        id: &ConcreteTypeId,
-    ) -> Result<(Type<'ctx>, Layout), super::error::Error>;
-}
-
-impl ProgramRegistryExt for ProgramRegistry<CoreType, CoreLibfunc> {
-    fn build_type<'ctx>(
-        &self,
-        context: &'ctx Context,
-        module: &Module<'ctx>,
-        registry: &ProgramRegistry<CoreType, CoreLibfunc>,
-        metadata: &mut MetadataStorage,
-        id: &ConcreteTypeId,
-    ) -> Result<Type<'ctx>, super::error::Error> {
-        registry
-            .get_type(id)?
-            .build(context, module, registry, metadata, id)
-    }
-
-    fn build_type_with_layout<'ctx>(
-        &self,
-        context: &'ctx Context,
-        module: &Module<'ctx>,
-        registry: &ProgramRegistry<CoreType, CoreLibfunc>,
-        metadata: &mut MetadataStorage,
-        id: &ConcreteTypeId,
-    ) -> Result<(Type<'ctx>, Layout), super::error::Error> {
-        let concrete_type = registry.get_type(id)?;
-
-        Ok((
-            concrete_type.build(context, module, registry, metadata, id)?,
-            concrete_type.layout(registry)?,
-        ))
-    }
-}
-
-pub trait RangeExt {
-    /// Width in bits when the offset is zero (aka. the natural representation).
-    fn zero_based_bit_width(&self) -> u32;
-    /// Width in bits when the offset is not necessarily zero (aka. the compact representation).
-    fn offset_bit_width(&self) -> u32;
-}
-
-impl RangeExt for Range {
-    fn zero_based_bit_width(&self) -> u32 {
-        // Formula for unsigned integers:
-        //     x.bits()
-        //
-        // Formula for signed values:
-        //   - Positive: (x.magnitude() + BigUint::one()).bits()
-        //   - Negative: (x.magnitude() - BigUint::one()).bits() + 1
-        //   - Zero: 0
-
-        let width = if self.lower.sign() == Sign::Minus {
-            let lower_width = (self.lower.magnitude() - BigUint::one()).bits() + 1;
-            let upper_width = {
-                let upper = &self.upper - &BigInt::one();
-                match upper.sign() {
-                    Sign::Minus => (upper.magnitude() - BigUint::one()).bits() + 1,
-                    Sign::NoSign => 0,
-                    Sign::Plus => (upper.magnitude() + BigUint::one()).bits(),
-                }
-            };
-
-            lower_width.max(upper_width) as u32
-        } else {
-            (&self.upper - &BigInt::one()).bits() as u32
-        };
-
-        // FIXME: Workaround for segfault in canonicalization (including LLVM 19).
-        width.max(1)
-    }
-
-    fn offset_bit_width(&self) -> u32 {
-        // FIXME: Workaround for segfault in canonicalization (including LLVM 19).
-        ((self.size() - BigInt::one()).bits() as u32).max(1)
-    }
 }
 
 #[cfg(test)]
